requirements.txt
munch~=2.5.0
requests~=2.27.1
blueink-client-python~=0.9.3